requirements.txt
amqp==5.1.1
aniso8601==9.0.1
apispec==6.3.0
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
CacheControl==0.13.1
cattrs==23.1.1
celery==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==4.4.2
exceptiongroup==1.1.1
Flask==0.10.1
flask-apispec==0.11.4
Flask-RESTful==0.3.9
gevent==22.10.2
greenlet==2.0.2
grequests==0.6.0
gunicorn==20.1.0
idna==3.4
imageio==2.31.0
imageio-ffmpeg==0.4.8
itsdangerous==0.24
Jinja2==2.7.2
kombu==5.3.0
librabbitmq==2.0.0
lxml==4.9.2
MarkupSafe==0.23
marshmallow==3.19.0
msgpack==1.0.5
numpy==1.24.3
opencv-python-headless==4.7.0.72
packaging==23.1
Pillow==9.5.0
platformdirs==3.5.1
proglog==0.1.10
prompt-toolkit==3.0.38
pycparser==2.21
pydub==0.25.1
PyJWT==2.7.0
python-dateutil==2.8.2
pytube==15.0.0
pytz==2023.3
pywebcopy==7.0.2
rabbitmq==0.2.0
redis==4.5.5
requests==2.30.0
requests-cache==1.0.1
six==1.16.0
soupsieve==2.4.1
tqdm==4.65.0
typing_extensions==4.6.2
tzdata==2023.3
url-normalize==1.4.3
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
webargs==8.2.0
Werkzeug==0.9.4
zope.event==4.6
zope.interface==6.0
